requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
requests>=2.26.0
matplotlib>=3.4.3
pydantic>=1.8.2
pydantic-settings==2.1.0
python-multipart==0.0.6 